Implement negation and disjunction (#18)
The argument "Every teacher is happy. Donald is not happy. Therefore, Donald is not a teacher." exemplifies modus tollens. We can implement this argumentative form by extending the DCG rules and question-answering engine to support negation, represented by a negation functor. The question-answering engine tries to prove "not P" by finding a clause "if P, then Q" and proving "not Q". Finally, the prove_question/2 predicate tries to prove both "P" and "not P" to handle both questions "Is Donald (not) happy?".

The argument “Every pixel is red or blue. Pixie is a pixel. Pixie is not blue. Therefore, Pixie is red.” is an example of exclusive disjunction. We implemented this for the binary case by extending the DCG rules and question-answering engine to support rules with disjunctive heads. If the clause to be proven appears in a disjunctive head, in either position, we try to prove the negation of the other disjunct. We have proven the original clause if the proof succeeds, assuming the disjuncts are exclusive. The prove_question/2 predicate further tries to prove "P" is true, "not P" is true, "not P" is false ("P" is true), and "P" is false ("not P") is true. This is ... not ideal.